[Refactor] 패키지명 변경으로 인한 수정
