+++ src/test/java/com/nhnacademy/bookpubshop/file/dummy/FileDummy.java
@@ -4,9 +4,8 @@ import com.nhnacademy.bookpubshop.coupontemplate.entity.CouponTemplate;
 import com.nhnacademy.bookpubshop.file.entity.File;
 import com.nhnacademy.bookpubshop.personalinquiry.entity.PersonalInquiry;
 import com.nhnacademy.bookpubshop.product.entity.Product;
-import com.nhnacademy.bookpubshop.customer_service.entity.CustomerService;
+import com.nhnacademy.bookpubshop.customersupport.entity.CustomerService;
 import com.nhnacademy.bookpubshop.review.entity.Review;
-import java.time.LocalDateTime;
 
 /**
  * Some description here.
